--- vehicle/saic/requests/api_config.go
@@ -1,8 +1,0 @@
-package requests
-
-const (
-	CONTENT_ENCRYPTED    = "1"
-	PARAM_AUTHENTICATION = "Basic c3dvcmQ6c3dvcmRfc2VjcmV0"
-	TENANT_ID            = "459771"
-	USER_TYPE            = "app"
-)
