+++ vehicle/saic/requests/types.go
@@ -1,5 +1,12 @@
 package requests
 
+const (
+	CONTENT_ENCRYPTED    = "1"
+	PARAM_AUTHENTICATION = "Basic c3dvcmQ6c3dvcmRfc2VjcmV0"
+	TENANT_ID            = "459771"
+	USER_TYPE            = "app"
+)
+
 type ChargeStatus struct {
 	RvsChargeStatus struct {
 		MileageSinceLastCharge    int
